added tests + models + edited main
