+++ backend_flask/app/__init__.py
@@ -24,5 +24,7 @@ def create_app(config_name):
 
     # attach routes and custom error pages here
 
+    from .main import main as main_blueprint
+    app.register_blueprint(main_blueprint)
 
     return app
