+++ backend_flask/app/main/views.py
@@ -1,0 +1,6 @@
+from flask import render_template
+from . import main
+
+@main.route('/')
+def index():
+    return render_template('index.html')
