+++ backend_flask/app/templates/index.html
@@ -1,0 +1,3 @@
+{% extends "base.html" %}
+
+<h1>Home Page</h1>
