+++ backend_flask/config.py
@@ -15,7 +15,6 @@ class Config:
 
 class DevelopmentConfig(Config):
     DEBUG = True
-    # ?? current name is 'data.sqlite, same as production ??'
     SQLALCHEMY_DATABASE_URI = os.environ.get('DEV_DATABASE_URL') or \
         'sqlite:///' + os.path.join(basedir, 'data-dev.sqlite')
