+++ backend_flask/requirements.txt
@@ -1,0 +1,92 @@
+alembic==1.6.5
+asgiref @ file:///tmp/build/80754af9/asgiref_1605055780383/work
+attrs==21.2.0
+Automat==20.2.0
+beautifulsoup4==4.9.3
+boto3 @ file:///tmp/build/80754af9/boto3_1613251195608/work
+botocore @ file:///tmp/build/80754af9/botocore_1613847367661/work
+brotlipy==0.7.0
+certifi==2020.12.5
+cffi @ file:///tmp/build/80754af9/cffi_1613246951236/work
+chardet==4.0.0
+click==8.0.1
+constantly==15.1.0
+cryptography @ file:///tmp/build/80754af9/cryptography_1613401443375/work
+cssselect==1.1.0
+dicttoxml==1.7.4
+dj-database-url==0.5.0
+Django @ file:///tmp/build/80754af9/django_1609784082548/work
+django-cors-headers==3.7.0
+django-crispy-forms==1.11.0
+django-filter==2.4.0
+django-heroku==0.3.1
+django-import-data==0.4.2
+django-storages==1.11.1
+djangorestframework==3.12.2
+dominate==2.6.0
+Flask==2.0.1
+Flask-Bootstrap==3.3.7.1
+Flask-Migrate==3.0.1
+Flask-SQLAlchemy==2.5.1
+gcloud==0.18.3
+googleapis-common-protos==1.52.0
+greenlet==1.1.0
+gunicorn==20.0.4
+h2==3.2.0
+hpack==3.0.0
+httplib2==0.19.0
+hyperframe==5.2.0
+hyperlink==21.0.0
+idna==2.10
+incremental==21.3.0
+itemadapter==0.2.0
+itemloaders==1.0.4
+itsdangerous==2.0.1
+Jinja2==3.0.1
+jmespath @ file:///tmp/build/80754af9/jmespath_1594304593830/work
+json2table==1.1.5
+jwcrypto==0.8
+lxml==4.6.3
+Mako==1.1.4
+Markdown==3.3.4
+MarkupSafe==2.0.1
+oauth2client==4.1.3
+olefile==0.46
+parsel==1.6.0
+Pillow @ file:///tmp/build/80754af9/pillow_1609786643840/work
+priority==1.3.0
+Protego==0.1.16
+protobuf==3.15.1
+psycopg2 @ file:///tmp/build/80754af9/psycopg2_1612298595717/work
+pyasn1==0.4.8
+pyasn1-modules==0.2.8
+pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
+pycryptodome==3.10.1
+PyDispatcher==2.0.5
+pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
+pyparsing==2.4.7
+Pyrebase4==4.4.3
+PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
+python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
+python-editor==1.0.4
+python-jwt==3.3.0
+pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
+queuelib==1.6.1
+requests==2.25.1
+requests-toolbelt==0.9.1
+rsa==4.7.1
+s3transfer @ file:///tmp/build/80754af9/s3transfer_1610404175646/work
+Scrapy==2.5.0
+service-identity==21.1.0
+six @ file:///tmp/build/80754af9/six_1605205306277/work
+soupsieve==2.2.1
+SQLAlchemy==1.4.19
+sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
+Twisted==21.2.0
+urllib3 @ file:///tmp/build/80754af9/urllib3_1611694770489/work
+visitor==0.1.3
+w3lib==1.22.0
+Werkzeug==2.0.1
+whitenoise==5.2.0
+xmltodict==0.12.0
+zope.interface==5.4.0
